worker/requirements.txt
kafka-python
python-dotenv
requests
psycopg2-binary
boto3
Django
